LayoutTests/media/video-seek-by-small-increment.html

<!DOCTYPE HTML>
<html>
    <head>
        <script src=media-file.js></script>
        <script src=video-test.js></script>

        <script>
            var seekedEventCount = 0;
            var seeksAttempted = 0;
            var increment = 0.0004;

            function seekIncrement()
            {
                // We want to verify that seeking by an increment smaller than the test movie's time scale
                // succeeds. test.mp4 has a time scale of 2500 (the smallest unit of time in that file
                // is 0.0004 seconds), so start with that and decrease by half each time.
                increment /= 2;
                return increment;
            }

            function seeked()
            {
                ++seekedEventCount;
            }

            function attemptSeek()
            {
                var now = video.currentTime;

                video.currentTime = now + seekIncrement();
                ++seeksAttempted;
                setTimeout(testAgain, 10);
            }

            function testAgain()
            {
                reportExpected(seeksAttempted == seekedEventCount, "seekedEventCount", "==", seeksAttempted, seekedEventCount);

                if (seeksAttempted == 10) {
                    endTest();
                    return;
                }

                setTimeout(attemptSeek, 100);
            }

            function start()
            {
                findMediaElement();
                video.src = findMediaFile("video", "content/test");

                waitForEvent('canplaythrough', testAgain);
                waitForEvent('seeked', seeked);
                waitForEvent('seeking');
                waitForEvent('play');
                waitForEvent('pause');
            }
        </script>
    </head>
    <body>
        <video controls></video>
        <p>Test seeking by very small increments.</p>
        <script>start()</script>
    </body>
</html>
