LayoutTests/media/video-seek-by-small-increment-expected.txt

Test seeking by very small increments.

EVENT(canplaythrough)
EXPECTED (seekedEventCount == '0') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '1') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '2') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '3') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '4') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '5') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '6') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '7') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '8') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '9') OK
EVENT(seeking)
EVENT(seeked)
EXPECTED (seekedEventCount == '10') OK
END OF TEST
